+++ Sources/TeslaSwift/Model/EnergySite.swift
@@ -21,12 +21,12 @@ open class EnergySite: Codable {
     open var resourceType: String
     open var siteName: String
     open var gatewayID: String
-    open var energyLeft: Int
-    open var totalPackEnergy: Int
-    open var percentageCharged: Int
+    open var energyLeft: Double
+    open var totalPackEnergy: Double
+    open var percentageCharged: Double
     open var batteryType: String
     open var backupCapable: Bool
-    open var batteryPower: Int
+    open var batteryPower: Double
     open var syncGridAlertEnabled: Bool
     open var breakerAlertEnabled: Bool
